+++ src/apis/artwork/getFollowedArtists.ts
@@ -5,7 +5,7 @@ import QUERY_KEYS from '@/constants/queryKeys';
 
 import { useQuery } from '@tanstack/react-query';
 
-export const getFollowedArtists = async () => {
+const getFollowedArtists = async () => {
   try {
     const response = await defaultClient.get(ARTWORK.followedArtists);
     return response.data.posts;
@@ -15,9 +15,11 @@ export const getFollowedArtists = async () => {
   }
 };
 
-export const useFollowedArtists = () => {
+const useFollowedArtists = () => {
   return useQuery({
     queryKey: [QUERY_KEYS.followedArtists],
     queryFn: getFollowedArtists,
   });
 };
+
+export default useFollowedArtists;
